App1
An edited form of the tutorial example. A scanner is used to get the user to input an integer inside a do-while statement, the desired value of which is 5. As long as the while input integer value is not equal to 5, the loop starts again. When "value" is equal to 5, the loop terminates, the scanner closes and a message is output indicating success. 

I have also encapsulated the entire do-while loop in a try/catch block to handle InputMismatchException, for instances where the user enters a non-integer during input. If this error is thrown, a message is output indicating an invalid value was entered.