App2
In this example program I simulated a test where the user has to correctly answer the question regarding the last name of the 1st US president to pass. I started by creating a scanner called "question". I then created two strings, one of which is the correct answer of "Washington", the other left as null to be defined later by user input. I then used a do-while statement that asks the test question, followed by the scanner which awaits the user input. Then an if statement is used to check if the user input & correct answer to the test were not of equal value, which I use the .equals() method to do. If they are not equal in value, a message is output saying telling the user to try again.

For the while part of the do-while statement, it checks to see if the user input and the correct answer are not equal while ignoring upper or lower case considerations. In this program, both "washington" and "Washington" are correct inputs.

Once the correct input is entered the scanner is closed and a message is output indicating 100%.